Proof of concept for Guava with SequencedCollections.
The main intent here is just to check that adding the SequencedCollection API
doesn't cause anything to fail. We make `ImmutableCollection` implement
`SequencedCollection`, which implies adding a bunch of `reversed()` methods in
descendant classes. Those methods are naively implemented. For production
quality we would want them to return views rather than constructing new
collections. That's not completely straightforward, though, because of
serialization.